% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_ariamsita.R
\name{theme_ariamsita}
\alias{theme_ariamsita}
\title{Theme Ariamsita}
\usage{
theme_ariamsita(...)
}
\description{
Some basic aesthetics for a minimal ggplot theme
}
\examples{
+theme_ariamsita
}
\keyword{ggplot,}
\keyword{theme}
